removing empty models, adding views, refactoring
