+++ vlogsApp/vlogsApp/ViewController/DetailScreen/DetailScreenViewController.swift
@@ -15,53 +15,28 @@ protocol DetailScreenViewControllerDelegate: AnyObject {
 
 class DetailScreenViewController: UIViewController {
     
+    var detailScreenView = DetailScreenView()
     var firebaseManager: FirebaseManager?
     weak var delegate: DetailScreenViewControllerDelegate?
-    var blog: Blog?
-    
-    lazy var titleLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = UIColor(named: "titleColor")
-        label.font = UIFont.boldSystemFont(ofSize: 18)
-        label.textAlignment = .center
-        label.numberOfLines = 0
-        label.text = blog?.title
-        return label
-    }()
-    
-    lazy var descriptionLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = UIColor(named: "titleColor")
-        label.font = UIFont.systemFont(ofSize: 14)
-        label.textAlignment = .center
-        label.numberOfLines = 0
-        label.text = blog?.description
-        return label
-    }()
-    
-    lazy var vlogImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.clipsToBounds = true
-        imageView.layer.cornerRadius = 20
-        imageView.layer.borderWidth = 1
-        imageView.contentMode = .scaleToFill
-        return imageView
-    }()
-    
-    lazy var vlogImage: UIImage = {
-        let image = UIImage()
-        return image
-    }()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        firebaseManager?.downloadPhoto(path: blog?.image ?? "", completion: { url in
-            self.vlogImageView.sd_setImage(with: url)
+        detailScreenView.delegate = self
+        
+        navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Edit", style: .done, target: self, action: #selector(editButtonTapped))
+        
+        view.addSubview(detailScreenView)
+        
+        detailScreenView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
+        }
+        
+        firebaseManager?.downloadPhoto(path: detailScreenView.blog?.image ?? "", completion: { url in
+            self.detailScreenView.vlogImageView.sd_setImage(with: url)
         })
         
         view.backgroundColor = UIColor(named: "backgroundColor")
-        configUI()
     }
     
     init(firebaseManager: FirebaseManager?) {
@@ -72,34 +47,11 @@ class DetailScreenViewController: UIViewController {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    func configUI() {
-        
-        navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Edit", style: .done, target: self, action: #selector(editButtonTapped))
+}
 
-        view.addSubview(titleLabel)
-        view.addSubview(descriptionLabel)
-        view.addSubview(vlogImageView)
-        
-        titleLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(vlogImageView.snp.bottom).offset(20)
-            make.leading.trailing.equalToSuperview().inset(12)
-        }
-        
-        descriptionLabel.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalTo(titleLabel.snp.bottom).offset(10)
-            make.leading.trailing.equalToSuperview().inset(12)
-        }
-        
-        vlogImageView.snp.makeConstraints { make in
-            make.centerX.equalToSuperview()
-            make.top.equalToSuperview().offset(20)
-            make.width.equalTo(350)
-            make.height.equalTo(250)
-        }
-    }
+//MARK: -
+
+extension DetailScreenViewController: DetailScreenProtocol {
     
     @objc func editButtonTapped() {
         
@@ -116,18 +68,18 @@ class DetailScreenViewController: UIViewController {
                let newDescription = alertController.textFields?[1].text {
                 
         
-                if let title = self.blog?.title, let description = self.blog?.description, let blogID = self.blog?.blogID {
+                if let title = self.detailScreenView.blog?.title, let description = self.detailScreenView.blog?.description, let blogID = self.detailScreenView.blog?.blogID {
                     self.firebaseManager?.updateBlogData(blogID: blogID, title: title, description: description, newTitle: newTitle, newDescription: newDescription, completion: { error in
                         if let error = error {
                             print("Error updating \(error)")
                         } else {
-                            self.blog?.title = newTitle
-                            self.blog?.description = newDescription
+                            self.detailScreenView.blog?.title = newTitle
+                            self.detailScreenView.blog?.description = newDescription
                             
-                            self.titleLabel.text = newTitle
-                            self.descriptionLabel.text = newDescription
+                            self.detailScreenView.titleLabel.text = newTitle
+                            self.detailScreenView.descriptionLabel.text = newDescription
                             
-                            guard let selectedBlog = self.blog else {
+                            guard let selectedBlog = self.detailScreenView.blog else {
                                 return
                             }
                             self.delegate?.didUpdateBlog(selectedBlog)
@@ -145,5 +97,5 @@ class DetailScreenViewController: UIViewController {
         
         present(alertController, animated: true, completion: nil)
     }
+    
 }
-
